requirements.txt
pandas
numpy
sentence-transformers
joblib
python-dateutil
haversine
torch
rapidfuzz